+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Event click for component frequency graph
 
 ### Changed
+- The action buttons in the component settings are stuck to the bottom edge of the modal window
 
 ### Fixed
 - change color of log`s icon when DS has finish with error