+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The action buttons in the component settings are stuck to the bottom edge of the modal window
 
 ### Fixed
+- Generation of component ID. The order of suffixes of the component ID suffixes added to the dashboard was incorrect in some cases.
 
 ## [1.150.0]