+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - cSvg component does not set the click token value.
 - Wrong text in the table when loading the dashboard "data is not displayed due to settings"
+- Table: Changing the token causes a reset to the first page.
 
 ## [1.150.0]