+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## [1.151.0]
+
+### Added
 - TableV2: handling "_columnOptions" in DS for setting columns options
 - TableV2: added new column types: riskAcc, riskFact, riskSum